+++ assets/javascripts/kitten/components/carousel/carousel/stories.js
@@ -81,6 +81,7 @@ storiesOf('Carousel/ProjectCarousel', module)
           CONTAINER_PADDING,
         )}
         paginationPosition={object('paginationPosition', paginationPosition)}
+        showPageSquares={boolean('showPageSquares', false)}
       >
         {data.map(item => (
           <CrowdfundingCard
@@ -219,7 +220,7 @@ storiesOf('Carousel/EngagementsCarousel', module)
       <StyledContainer fullWidthBelowScreenSize="S">
         <Carousel
           itemMinWidth={110}
-          hidePagination
+          hidePaginationOnMobile
           baseItemMarginBetween={20}
           paginationPosition={{ default: 'right' }}
           showOtherPages={boolean('Show other pages?', false)}